refactor: Move controllers and middleware into files
This extracts the individual controllers and middlewares into their own
files. This should improve readability of the code and matches the
pattern set out by the other larger apps.
