+++ app/person/controllers/image.test.js
@@ -3,19 +3,13 @@ const axios = require('axios')
 const personService = {
   getImageUrl: sinon.stub(),
 }
-const presenters = require('../../common/presenters')
-const nunjucksGlobals = require('../../config/nunjucks/globals')
+const nunjucksGlobals = require('../../../config/nunjucks/globals')
 
-const controllers = require('./controllers')
+const image = require('./image')
 
 const mockPersonId = 'f6e1f57c-7d07-41d2-afee-1662f5103e2d'
 const placeholderImage = '/images/person-placeholder.png'
 const manifestImagePath = '/images/person-placeholder.19s41n.png'
-
-const mockPerson = {
-  _fullname: 'DOE, JOHN',
-  id: mockPersonId,
-}
 
 describe('Person app', function () {
   describe('Controllers', function () {
@@ -43,7 +37,7 @@ describe('Person app', function () {
       context('when person service errors', function () {
         beforeEach(async function () {
           personService.getImageUrl.rejects(new Error())
-          await controllers.image(placeholderImage)(req, res)
+          await image(placeholderImage)(req, res)
         })
 
         it('should not call res.end', function () {
@@ -71,7 +65,7 @@ describe('Person app', function () {
         context('when axios errors', function () {
           beforeEach(async function () {
             axios.get.rejects(new Error())
-            await controllers.image(placeholderImage)(req, res)
+            await image(placeholderImage)(req, res)
           })
 
           it('should not call res.end', function () {
@@ -101,7 +95,7 @@ describe('Person app', function () {
 
           beforeEach(async function () {
             axios.get.resolves(mockResponse)
-            await controllers.image(placeholderImage)(req, res)
+            await image(placeholderImage)(req, res)
           })
 
           it('should set headers', function () {
@@ -120,79 +114,6 @@ describe('Person app', function () {
           it('should not redirect to placeholder images', function () {
             expect(res.redirect).not.to.be.called
           })
-        })
-      })
-    })
-
-    describe('#personalDetails()', function () {
-      let req, res
-
-      beforeEach(function () {
-        sinon.stub(presenters, 'personToSummaryListComponent').returnsArg(0)
-        req = {
-          person: mockPerson,
-          t: sinon.stub().returnsArg(0),
-        }
-        res = {
-          breadcrumb: sinon.stub().returnsThis(),
-          render: sinon.stub(),
-        }
-
-        controllers.personalDetails(req, res)
-      })
-
-      it('should set breadcrumb', function () {
-        expect(req.t).to.be.called
-        expect(res.breadcrumb).to.be.calledOnceWithExactly({
-          text: 'person::personal_details.heading',
-        })
-      })
-
-      it('should render correct template', function () {
-        expect(res.render.args[0][0]).to.equal('person/views/personal-details')
-      })
-
-      describe('locals', function () {
-        let locals
-
-        beforeEach(function () {
-          locals = res.render.args[0][1]
-        })
-
-        it('should pass correct keys', function () {
-          expect(locals).to.have.all.keys([
-            'fullname',
-            'identityBar',
-            'personalDetailsSummary',
-          ])
-        })
-
-        it('should set fullname', function () {
-          expect(locals.fullname).to.equal('DOE, JOHN')
-        })
-
-        it('should set identityBar', function () {
-          expect(locals.identityBar).to.deep.equal({
-            classes: 'sticky',
-            caption: {
-              text: 'person::page_caption',
-            },
-            heading: {
-              html: 'person::page_heading',
-            },
-          })
-
-          expect(req.t).to.have.been.calledWithExactly('person::page_caption')
-          expect(req.t).to.have.been.calledWithExactly('person::page_heading', {
-            name: 'DOE, JOHN',
-          })
-        })
-
-        it('should set personal details summary', function () {
-          expect(locals.personalDetailsSummary).to.deep.equal(mockPerson)
-          expect(
-            presenters.personToSummaryListComponent
-          ).to.have.been.calledOnceWithExactly(mockPerson)
         })
       })
     })
