+++ app/person/controllers/index.js
@@ -1,0 +1,4 @@
+const image = require('./image')
+const personalDetails = require('./personal-details')
+
+module.exports = { image, personalDetails }
