+++ app/person/middleware/index.js
@@ -1,0 +1,4 @@
+const setBreadcrumb = require('./set-breadcrumb')
+const setPerson = require('./set-person')
+
+module.exports = { setBreadcrumb, setPerson }
