+++ app/person/middleware/set-breadcrumb.js
@@ -1,0 +1,12 @@
+module.exports = (req, res, next) => {
+  const { _fullname: name } = req.person || {}
+
+  if (name) {
+    res.breadcrumb({
+      text: name,
+      href: req.baseUrl,
+    })
+  }
+
+  next()
+}
