+++ app/person/middleware/set-person.js
@@ -1,0 +1,14 @@
+module.exports = async (req, res, next) => {
+  const personId = req.params.personId
+
+  if (!personId) {
+    return next()
+  }
+
+  try {
+    req.person = await req.services.person.getById(personId)
+    next()
+  } catch (error) {
+    next(error)
+  }
+}
